# backend/src/openapi.yaml
openapi: 3.1.0
info:
  title: PatchIQ API
  description: |
    PatchIQ - Enterprise Patch Management System API

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Agent API
    The Agent API (`/api/agent/*`) is used by PatchIQ agents to communicate with the backend.
    These endpoints use agent-specific authentication via X-Agent-Id and X-Agent-Version headers.
  version: 1.0.0
  contact:
    name: PatchIQ Support
    email: support@patchiq.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3007
    description: Local Development
  - url: https://api.patchiq.io
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /v1/auth/login
    agentAuth:
      type: apiKey
      in: header
      name: X-Agent-Id
      description: Agent identifier for agent API endpoints
  schemas:
...
    Patch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        version:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [available, testing, approved, deployed, deprecated]
        releaseDate:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
paths:
...
  /v1/patches:
    get:
      summary: List patches
      description: Get a paginated list of patches
      tags:
        - Patches
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
      responses:
        '200':
          description: List of patches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Patch'
  /v1/patches/{id}:
    get:
      summary: Get patch
      description: Get a specific patch by ID
      tags:
        - Patches
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Patch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patch'

  # Vulnerabilities endpoints
...